requirements.txt
loguru==0.7.2
inquirer==3.2.1
termcolor==2.4.0
better-automation==2.0.2
web3==6.14.0
aiofiles==23.2.1
pillow==10.2.0
fake-useragent==1.4.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
Faker==22.4.0
pytz==2023.3.post1
aiohttp_proxy==0.1.2
aiohttp-socks==0.8.0
